requirements.txt
flask==1.1.1
torch==1.4.0
torchvision==0.5.0
scikit-learn==0.22.2
numpy==1.18.2
gunicorn==20.0.4
pillow==7.0.0
tensorboard==2.1.1
matplotlib==3.2.1
setuptools==41.0.0
